requirements.txt
python-dotenv==1.0.0
aiohttp==3.9.3
requests==2.32.3
pyyaml==6.0.1
tenacity==8.2.3